Updated app.js
1) Updated getInitial query to include a join to pull the appointment
information, as well as the patient information in a single query. This
query also discards the id of the patient, as it had a conflicting name
with the id of the appointment, and the patient id is already stored in
the appointment table as "patient".

2) Added a new query "getNewAppointmentPatient" that gets called after
the updateRecords query. Currently, updateRecords listens for changes in
the appointments table. When an existing appointment's patient changes
from null to a primary key value, this triggers the
"getNewAppointmentPatient" query, which returns the newly assinged
patients information.