Updated index.html
A lot of changes here.. Sorry!

1) The rendering of the dynamic portion of the page has been moved from
the body of the HTML to a template(At the bottom of the page for now, it
will be imported later). This allows each instance of the appointment to
rendered as a seperate element that looks like the template.

2) Added comments everywhere

3) Added a Vue Component to handle passing data to the template through
the prop ['item']

4) Updated the initialRecords function to add all of the patient info
and the appointment info, if the patient isn't null.

5)Updated the possible conditions for the change listener function,
updateRecordsResults, to include additons, deletions, and updates of
appointments.

6)When an appointment is updated, and the patient field is no longer
null, the getNewAppointmentPatient function is called, which will return
the newly assigned patients information. This then gets added to the
local data, and rendered dynamically by Vue

7) The template itself uses the prop ['item'], which is a way of passing
information into the template. In this case [item] is the joined patient
and appointment information. It then renders this information using tags
like {{ item.name }}

8) The dissmiss button in the template deletes the whole panel element(
which holds all the appointment info) from the page.